Add new resource definition for Azure Virtual Network
This commit introduces a new Terraform configuration file, resource.azurerm_virtual_network.vnet.tf, which defines an Azure Virtual Network (VNET) resource. The VNET is configured with a name based on the resource group name, a specified location, and an address space of 10.0.0.0/16. This addition enhances the infrastructure setup for the project.